services/room/pyproject.toml
[project]
name = "room"
version = "0.1.0"
description = "Room collaboration gRPC service"
requires-python = ">=3.11"
dependencies = [
    "pb",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "pydantic>=2.10.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
pb = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/room"]
